fixing scroll view in information sheet
