+++ bubble-jam/Source/Entity/AudioAndPropeties.swift
@@ -17,7 +17,7 @@ struct AudioAndPropeties: Equatable {
 struct AudioDetails: Equatable {
     let notes: [Note]?
     let description: String?
-    let rules: [String]
+    var rules: [String]
     let bpm: Int
 }
