--- bubble-jam/Source/Presentation/View/Bubblegum/Components/Sheet/Components/InformationsScrollView.swift
@@ -1,40 +1,0 @@
-//
-//  InformationsScrollView.swift
-//  bubble-jam
-//
-//  Created by Thiago Henrique on 27/02/23.
-//
-
-import UIKit
-
-class InformationsScrollView: UIView {
-    
-    private let scrollView: UIScrollView = {
-        let scrollView = UIScrollView()
-        scrollView.translatesAutoresizingMaskIntoConstraints = false
-        scrollView.isScrollEnabled = true
-        return scrollView
-    }()
-    
-    override init(frame: CGRect) {
-        super.init(frame: .zero)
-        buildLayout()
-    }
-    
-    required init?(coder: NSCoder) { nil }
-}
-
-extension InformationsScrollView: ViewCoding {
-    func setupHierarchy() {
-        self.addSubview(scrollView)
-    }
-    
-    func setupConstraints() {
-        NSLayoutConstraint.activate([
-            scrollView.leadingAnchor.constraint(equalTo: self.leadingAnchor),
-            scrollView.topAnchor.constraint(equalTo: self.topAnchor),
-            scrollView.trailingAnchor.constraint(equalTo: self.trailingAnchor),
-            scrollView.bottomAnchor.constraint(equalTo: self.bottomAnchor)
-        ])
-    }
-}
